! pull some (most) items out of theme and back into the abstract custom themes can still override, most never will so this lightens the theme.php that may be included by a custom theme.
